refactor(kit): collapse the runtime/ split into one module per actor
aether-kit no longer mirrors the identity/runtime two-layer split from the
native crates: the `runtime` feature and the `runtime/` subtree are gone, and
each system is one module under the crate root that co-locates the actor with
its own `kinds` file. `lib.rs` keeps only the shared consts, the `export!`
pair, and the root re-exports, so the public kind paths are unchanged. The
substrate-bundle dev-dep drops `default-features = false`.

The world module additionally splits its wire kinds out of the data layer
(`world/kinds.rs`), and the locomotion actor's pathfinding moves into
`locomotion/path.rs`. The compositing widget and the reference panel share
their `flush_membership` / child-list-accept logic through `widget`
free functions instead of duplicating it.

Qodana: the `locomotion` module (being phased out to legacy) is excluded from
the DuplicatedCode inspection, since its findings pair pre-existing
camera/movement code shared with the mover actor.
