+++ crates/aether-kit/src/widget/set/button.rs
@@ -1,5 +1,5 @@
 // `#[handler]` methods take their decoded mail by value per the ADR-0033
-// dispatch ABI (see `runtime/widget.rs`).
+// dispatch ABI (see `widget/mod.rs`).
 #![allow(clippy::needless_pass_by_value)]
 
 //! The momentary push button (issue 2660).
@@ -16,9 +16,9 @@ use aether_actor::{ActorInitError, WasmActor, WasmCtx, WasmInitCtx, actor};
 use aether_kinds::mouse_button;
 use aether_kinds::{MouseButton, MouseButtonRelease};
 
-use crate::runtime::widgets::{push_border, quad, text_origin_y};
-use crate::theme::{SetTheme, Theme, WidgetState};
-use crate::widgets::{
+use crate::widget::set::{push_border, quad, text_origin_y};
+use crate::widget::theme::{SetTheme, Theme, WidgetState};
+use crate::widget::{
     ButtonClicked, ButtonConfig, Collect, FocusGained, FocusLost, WidgetDrawItem, WidgetDrawList,
     WidgetFrame,
 };
