+++ crates/aether-kit/src/widget/set/slider.rs
@@ -1,5 +1,5 @@
 // `#[handler]` methods take their decoded mail by value per the ADR-0033
-// dispatch ABI (see `runtime/widget.rs`).
+// dispatch ABI (see `widget/mod.rs`).
 #![allow(clippy::needless_pass_by_value)]
 
 //! The horizontal value slider (issue 2660).
@@ -18,9 +18,9 @@ use aether_kinds::keycode::{KEY_DOWN, KEY_LEFT, KEY_RIGHT, KEY_UP};
 use aether_kinds::mouse_button;
 use aether_kinds::{Key, MouseButton, MouseButtonRelease, MouseMove};
 
-use crate::runtime::widgets::{push_border, quad};
-use crate::theme::{SetTheme, Theme, WidgetState};
-use crate::widgets::{
+use crate::widget::set::{push_border, quad};
+use crate::widget::theme::{SetTheme, Theme, WidgetState};
+use crate::widget::{
     Collect, FocusGained, FocusLost, SliderChanged, SliderConfig, WidgetDrawItem, WidgetDrawList,
     WidgetFrame,
 };
